Change Management Calendar
This JavaScript/jQuery calendar was built to display IT change requests in a useful format, in the interest of improving change management processes and planning. The idea is that a technician, and/or change manager, can see: what changes have occurred in the past few days; what changes are planned; and what change requests are queued up for review. See Readme for more details, and in-line comments in .js file.